EPMCUXDA-6947 move entities common interfaces up and add new interface
