+++ src/entities/common-interfaces.ts
@@ -3,3 +3,8 @@ export interface Dictionary {
   desc_rus?: string | null;
   desc_byn?: string | null;
 }
+
+export interface AbleToExportAndImportEuring {
+  exportEURING(): string;
+  importEURING(code: string): any;
+}
